WIP: Read from Database table for new updates
We are using Postgres as an event queue to asynchronously process
received updates, without care about its order.

The strategy adopted when consuming from postgres is to use a DELETE
operation to delete and receive N events, and process it. If it fails,
the process will try to place it again in the Database table. If the
process crashes, the Supervisor will retrieve the process state and try
to save it into the database himself.